【量子-内测后台】财务管理-入款配置-第三方APP支付中; 禁用功能未生效

+++ src/main/java/com/lzkj/mobile/controller/ThirdAppPayController.java
@@ -30,7 +30,7 @@ public class ThirdAppPayController {
         List<ThirdAppPayConfigVO> configListRet = new ArrayList<>();
         // 过滤掉不可用数据
         configList.stream().forEach(obj -> {
-            if(new Integer(1).equals(obj.getEnableState())){
+            if(new Integer(0).equals(obj.getEnableState())){
                 configListRet.add(obj);
             }
         });
